Hosts support HTTP health check
