+++ pkg/microservice/aslan/core/system/service/private_key.go
@@ -28,6 +28,7 @@ import (
 	"github.com/koderover/zadig/pkg/setting"
 	"github.com/koderover/zadig/pkg/tool/crypto"
 	e "github.com/koderover/zadig/pkg/tool/errors"
+	"github.com/koderover/zadig/pkg/types"
 )
 
 func ListPrivateKeys(encryptedKey string, log *zap.SugaredLogger) ([]*commonmodels.PrivateKey, error) {
@@ -42,7 +43,7 @@ func ListPrivateKeys(encryptedKey string, log *zap.SugaredLogger) ([]*commonmode
 	}
 	for _, key := range resp {
 		if key.Probe == nil {
-			key.Probe = &commonmodels.Probe{ProbeScheme: setting.ProtocolTCP}
+			key.Probe = &types.Probe{ProbeScheme: setting.ProtocolTCP}
 		}
 		key.PrivateKey, err = crypto.AesEncryptByKey(key.PrivateKey, aesKey.PlainText)
 		if err != nil {
