+++ pkg/microservice/cron/core/service/scheduler/schedule_env.go
@@ -35,6 +35,7 @@ import (
 	"github.com/koderover/zadig/pkg/microservice/cron/core/service"
 	"github.com/koderover/zadig/pkg/microservice/cron/core/service/client"
 	"github.com/koderover/zadig/pkg/setting"
+	"github.com/koderover/zadig/pkg/types"
 )
 
 // UpsertEnvServiceScheduler ...
@@ -205,7 +206,7 @@ func runProbe(healthCheck *service.PmHealthCheck, address string, log *zap.Sugar
 	timeout := time.Duration(healthCheck.TimeOut) * time.Second
 	switch healthCheck.Protocol {
 	case setting.ProtocolHTTP, setting.ProtocolHTTPS:
-		if message, err = doHTTPProbe(healthCheck.Protocol, address, healthCheck.Path, healthCheck.Port, []*service.HTTPHeader{}, timeout, "", log); err != nil {
+		if message, err = doHTTPProbe(healthCheck.Protocol, address, healthCheck.Path, healthCheck.Port, []*types.HTTPHeader{}, timeout, "", log); err != nil {
 			log.Errorf("doHttpProbe err:%v", err)
 			return Failure, err
 		}
@@ -240,7 +241,7 @@ func doTCPProbe(addr string, port int, timeout time.Duration, log *zap.SugaredLo
 	return Success, nil
 }
 
-func doHTTPProbe(protocol, address, path string, port int, headerList []*service.HTTPHeader, timeout time.Duration, responseSuccessFlag string, log *zap.SugaredLogger) (string, error) {
+func doHTTPProbe(protocol, address, path string, port int, headerList []*types.HTTPHeader, timeout time.Duration, responseSuccessFlag string, log *zap.SugaredLogger) (string, error) {
 	tlsConfig := &tls.Config{InsecureSkipVerify: true}
 	transport := &http.Transport{
 		TLSClientConfig:   tlsConfig,
@@ -281,7 +282,7 @@ func doHTTPProbe(protocol, address, path string, port int, headerList []*service
 		log.Infof("Probe succeeded for %s, Response: %v", url, *res)
 		return Success, nil
 	}
-	log.Infof("Probe failed for %s, response body: %v", url, string(body))
+	log.Warnf("Probe failed for %s, response body: %v", url, string(body))
 	return Failure, fmt.Errorf("HTTP probe failed with statuscode: %d", res.StatusCode)
 }
 
@@ -318,7 +319,7 @@ func formatURL(protocol, address, path string, port int) (string, error) {
 	return fmt.Sprintf("%s://%s:%d/%s", protocol, address, port, path), nil
 }
 
-func buildHeader(headerList []*service.HTTPHeader) http.Header {
+func buildHeader(headerList []*types.HTTPHeader) http.Header {
 	headers := make(http.Header)
 	for _, header := range headerList {
 		headers[header.Name] = append(headers[header.Name], header.Value)
